What I have been working on.
The work is still in progress, but I was able to get a working 3D graph to print the data. The csv files hold that data.
Both programs run using a python compiler...
sampleGroup.py is used to add more data to the csv files.
Snappetite.py is used to test the image files (basically to send back information).